Everything changed
Still in Perl, but everything else is different. Note that "echo" effects have been disabled - though they run great on most Linuxes, the weakened PCM of the Pi usually chokes on bends and echos at once. If you're running it on a faster Linux, consider un-commenting that part of the syntax.  It's also "tuned-down" a bit, to suit "DroneTV" - a live stream on YouTube serving this 24/7.